Separate tailwind config into two parts: static and dynamic
Tailwind has to use JS-based config, since plugins are functions and therefore can not be defined in a JSON file. However this makes the config not editable from projenrc file. In order to make the static part of the config editable it is moved into a separate JSON file (created by projen JSONFile class) and thus can be edited via its addOverride method.